fix(create): if importing and catch don't delete deployment

+++ controllers/create.deployment.controller.js
@@ -207,7 +207,7 @@ router.post(['/', '/import'], async (req, res, next) => {
       })
   } catch (error) {
     // console.log(error)
-    if (deploymentId) {
+    if (deploymentId && !importing) {
       await Deployment.findByIdAndDelete(deploymentId)
     }
     next(error)
